hybrid_agent_exploration/requirements.txt
rdkit-pypi>=2023.03.1
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2
optuna>=3.4.0
shap>=0.43.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
omegaconf>=2.3.0
rich>=13.0.0
autogluon>=1.0.0
openpyxl>=3.1.0
umap-learn>=0.5.3
mordred>=1.2.0
selfies>=2.1.0
